tests/wpt/web-platform-tests/workers/WorkerLocation_pathname.htm

<!DOCTYPE html>
<title> WorkerLocation URL decomposition IDL attribute: pathname </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/WorkerLocation.js');
  worker.onmessage = t.step_func_done(function(e) {
    var pathname = location.pathname;
    var expected = pathname.substring(0, pathname.lastIndexOf('/')) + '/support/WorkerLocation.js';
    assert_equals(e.data.pathname, expected);
  });
});
</script>
